+++ src/axom/quest/detail/MeshTester_detail.hpp
@@ -341,27 +341,16 @@ struct CandidateFinder<AccelType::BVH, ExecSpace, FloatType>
     counts.resize(this->m_aabbs.size());
 
     // Search for intersecting bounding boxes of triangles
-    IndexType* candidatesData = nullptr;
-    IndexType ncandidates = bvh.findBoundingBoxes(offsets.data(),
-                                                  counts.data(),
-                                                  candidatesData,
-                                                  this->m_aabbs.size(),
-                                                  this->m_aabbs.view());
+    bvh.findBoundingBoxes(offsets,
+                          counts,
+                          m_currCandidates,
+                          this->m_aabbs.size(),
+                          this->m_aabbs.view());
 
-    m_currCandidates.reset(candidatesData);
-    return axom::ArrayView<IndexType, 1, Space>(candidatesData, ncandidates);
+    return m_currCandidates;
   }
 
-  // Helper functor used by unique_ptr for deleting the result array returned
-  // from the BVH query
-  struct CandidateDeleter
-  {
-    void operator()(IndexType* ptr) const { axom::deallocate(ptr); }
-  };
-
-  std::unique_ptr<IndexType, CandidateDeleter> m_currCandidates {
-    nullptr,
-    CandidateDeleter {}};
+  axom::Array<IndexType> m_currCandidates;
 };
 
 /*!
@@ -439,15 +428,12 @@ struct CandidateFinder<AccelType::ImplicitGrid, ExecSpace, FloatType>
       this->m_aabbs.size(),
       allocatorId);
     gridIndex.insert(this->m_aabbs.size(), this->m_aabbs.data());
-    axom::Array<IndexType> offsetsTmp, countsTmp;
-    gridIndex.getCandidatesAsArray(this->m_aabbs.size(),
-                                   this->m_aabbs.data(),
-                                   offsetsTmp,
-                                   countsTmp,
-                                   m_currCandidates);
 
-    offsets = offsetsTmp;
-    counts = countsTmp;
+    offsets.resize(this->m_aabbs.size());
+    counts.resize(this->m_aabbs.size());
+
+    gridIndex.getCandidatesAsArray(this->m_aabbs, offsets, counts, m_currCandidates);
+
     return m_currCandidates;
   }