+++ src/axom/spin/tests/spin_bvh.cpp
@@ -345,11 +345,10 @@ void check_find_bounding_boxes3d()
   }
 
   // traverse the BVH to find the candidates for all the bounding boxes
-  IndexType* offsets = axom::allocate<IndexType>(N);
-  IndexType* counts = axom::allocate<IndexType>(N);
-  IndexType* candidates = nullptr;
+  axom::Array<IndexType> offsets(N);
+  axom::Array<IndexType> counts(N);
+  axom::Array<IndexType> candidates;
   bvh.findBoundingBoxes(offsets, counts, candidates, N, query_boxes);
-  EXPECT_TRUE(candidates != nullptr);
 
   // flag cells that are found by the bounding box ID
   int* iblank = mesh.createField<int>("iblank", mint::CELL_CENTERED);
@@ -405,9 +404,6 @@ void check_find_bounding_boxes3d()
   }  // END for all mesh cells
 
   // deallocate
-  axom::deallocate(offsets);
-  axom::deallocate(candidates);
-  axom::deallocate(counts);
   axom::deallocate(aabbs);
 
   axom::deallocate(query_boxes);
@@ -466,11 +462,10 @@ void check_find_bounding_boxes2d()
   }
 
   // traverse the BVH to find the candidates for all the bounding boxes
-  IndexType* offsets = axom::allocate<IndexType>(N);
-  IndexType* counts = axom::allocate<IndexType>(N);
-  IndexType* candidates = nullptr;
+  axom::Array<IndexType> offsets(N);
+  axom::Array<IndexType> counts(N);
+  axom::Array<IndexType> candidates;
   bvh.findBoundingBoxes(offsets, counts, candidates, N, query_boxes);
-  EXPECT_TRUE(candidates != nullptr);
 
   // flag cells that are found by the bounding box ID
   int* iblank = mesh.createField<int>("iblank", mint::CELL_CENTERED);
@@ -510,9 +505,6 @@ void check_find_bounding_boxes2d()
   }  // END for all mesh cells
 
   // deallocate
-  axom::deallocate(offsets);
-  axom::deallocate(candidates);
-  axom::deallocate(counts);
   axom::deallocate(aabbs);
 
   axom::deallocate(query_boxes);
@@ -574,11 +566,10 @@ void check_find_rays3d()
   }
 
   // traverse the BVH to find the candidates for all the centroids
-  IndexType* offsets = axom::allocate<IndexType>(N);
-  IndexType* counts = axom::allocate<IndexType>(N);
-  IndexType* candidates = nullptr;
+  axom::Array<IndexType> offsets(N);
+  axom::Array<IndexType> counts(N);
+  axom::Array<IndexType> candidates;
   bvh.findRays(offsets, counts, candidates, N, query_rays);
-  EXPECT_TRUE(candidates != nullptr);
 
   // flag cells that are found by the ray ID
   int* iblank = mesh.createField<int>("iblank", mint::CELL_CENTERED);
@@ -635,9 +626,6 @@ void check_find_rays3d()
   }  // END for all mesh cells
 
   // deallocate
-  axom::deallocate(offsets);
-  axom::deallocate(candidates);
-  axom::deallocate(counts);
   axom::deallocate(aabbs);
 
   axom::deallocate(query_rays);
@@ -708,11 +696,10 @@ void check_find_rays2d()
   }
 
   // traverse the BVH to find the candidates for all the centroids
-  IndexType* offsets = axom::allocate<IndexType>(N);
-  IndexType* counts = axom::allocate<IndexType>(N);
-  IndexType* candidates = nullptr;
+  axom::Array<IndexType> offsets(N);
+  axom::Array<IndexType> counts(N);
+  axom::Array<IndexType> candidates;
   bvh.findRays(offsets, counts, candidates, N, query_rays);
-  EXPECT_TRUE(candidates != nullptr);
 
   // flag cells that are found by the ray ID
   int* iblank = mesh.createField<int>("iblank", mint::CELL_CENTERED);
@@ -756,9 +743,6 @@ void check_find_rays2d()
   }  // END for all mesh cells
 
   // deallocate
-  axom::deallocate(offsets);
-  axom::deallocate(candidates);
-  axom::deallocate(counts);
   axom::deallocate(aabbs);
 
   axom::deallocate(query_rays);
@@ -822,12 +806,10 @@ void check_find_points3d()
   }
 
   // traverse the BVH to find the candidates for all the centroids
-  IndexType* offsets = axom::allocate<IndexType>(ncells);
-  IndexType* counts = axom::allocate<IndexType>(ncells);
-  IndexType* candidates = nullptr;
+  axom::Array<IndexType> offsets(ncells);
+  axom::Array<IndexType> counts(ncells);
+  axom::Array<IndexType> candidates;
   bvh.findPoints(offsets, counts, candidates, ncells, centroids);
-
-  EXPECT_TRUE(candidates != nullptr);
 
   spin::UniformGrid<IndexType, NDIMS> ug(lo, hi, res);
 
@@ -838,8 +820,6 @@ void check_find_points3d()
     EXPECT_EQ(counts[i], 1);
     EXPECT_EQ(donorCellIdx, candidates[offsets[i]]);
   }  // END for all cell centroids
-
-  axom::deallocate(candidates);
 
   // check points that are outside by shifting the query points
   constexpr double OFFSET = 10.0;
@@ -857,9 +837,6 @@ void check_find_points3d()
     EXPECT_EQ(counts[i], 0);
   }
 
-  axom::deallocate(offsets);
-  axom::deallocate(candidates);
-  axom::deallocate(counts);
   axom::deallocate(aabbs);
 
   axom::setDefaultAllocator(current_allocator);
@@ -921,12 +898,10 @@ void check_find_points2d()
   }
 
   // traverse the BVH to find the candidates for all the centroids
-  IndexType* offsets = axom::allocate<IndexType>(ncells);
-  IndexType* counts = axom::allocate<IndexType>(ncells);
-  IndexType* candidates = nullptr;
+  axom::Array<IndexType> offsets(ncells);
+  axom::Array<IndexType> counts(ncells);
+  axom::Array<IndexType> candidates;
   bvh.findPoints(offsets, counts, candidates, ncells, centroids);
-
-  EXPECT_TRUE(candidates != nullptr);
 
   spin::UniformGrid<IndexType, NDIMS> ug(lo, hi, res);
 
@@ -937,8 +912,6 @@ void check_find_points2d()
     EXPECT_EQ(counts[i], 1);
     EXPECT_EQ(donorCellIdx, candidates[offsets[i]]);
   }  // END for all cell centroids
-
-  axom::deallocate(candidates);
 
   // check points that are outside by shifting the query points
   constexpr double OFFSET = 10.0;
@@ -955,9 +928,6 @@ void check_find_points2d()
     EXPECT_EQ(counts[i], 0);
   }
 
-  axom::deallocate(offsets);
-  axom::deallocate(candidates);
-  axom::deallocate(counts);
   axom::deallocate(aabbs);
 
   axom::setDefaultAllocator(current_allocator);
@@ -1009,14 +979,12 @@ void check_single_box2d()
   PointType* centroid_device = axom::allocate<PointType>(1);
   axom::copy(centroid_device, &centroid, sizeof(PointType));
 
-  IndexType* offsets = axom::allocate<IndexType>(NUM_BOXES);
-  IndexType* counts = axom::allocate<IndexType>(NUM_BOXES);
-  IndexType* candidates = nullptr;
+  axom::Array<IndexType> offsets(NUM_BOXES);
+  axom::Array<IndexType> counts(NUM_BOXES);
+  axom::Array<IndexType> candidates;
   bvh.findPoints(offsets, counts, candidates, NUM_BOXES, centroid_device);
-  EXPECT_TRUE(candidates != nullptr);
   EXPECT_EQ(counts[0], 1);
   EXPECT_EQ(0, candidates[offsets[0]]);
-  axom::deallocate(candidates);
 
   // shift centroid outside of the BVH, should return no candidates.
   centroid[0] += 10.0;
@@ -1027,9 +995,6 @@ void check_single_box2d()
 
   axom::deallocate(centroid_device);
   axom::deallocate(boxes);
-  axom::deallocate(offsets);
-  axom::deallocate(counts);
-  axom::deallocate(candidates);
   axom::setDefaultAllocator(current_allocator);
 }
 
@@ -1078,14 +1043,12 @@ void check_single_box3d()
   PointType* centroid_device = axom::allocate<PointType>(1);
   axom::copy(centroid_device, &centroid, sizeof(PointType));
 
-  IndexType* offsets = axom::allocate<IndexType>(NUM_BOXES);
-  IndexType* counts = axom::allocate<IndexType>(NUM_BOXES);
-  IndexType* candidates = nullptr;
+  axom::Array<IndexType> offsets(NUM_BOXES);
+  axom::Array<IndexType> counts(NUM_BOXES);
+  axom::Array<IndexType> candidates;
   bvh.findPoints(offsets, counts, candidates, NUM_BOXES, centroid_device);
-  EXPECT_TRUE(candidates != nullptr);
   EXPECT_EQ(counts[0], 1);
   EXPECT_EQ(0, candidates[offsets[0]]);
-  axom::deallocate(candidates);
 
   // shift centroid outside of the BVH, should return no candidates.
   centroid[0] += 10.0;
@@ -1096,9 +1059,6 @@ void check_single_box3d()
 
   axom::deallocate(centroid_device);
   axom::deallocate(boxes);
-  axom::deallocate(offsets);
-  axom::deallocate(counts);
-  axom::deallocate(candidates);
   axom::setDefaultAllocator(current_allocator);
 }
 
@@ -1231,12 +1191,10 @@ void check_find_points_zip3d()
   }
 
   // traverse the BVH to find the candidates for all the centroids
-  IndexType* offsets = axom::allocate<IndexType>(ncells);
-  IndexType* counts = axom::allocate<IndexType>(ncells);
-  IndexType* candidates = nullptr;
+  axom::Array<IndexType> offsets(ncells);
+  axom::Array<IndexType> counts(ncells);
+  axom::Array<IndexType> candidates;
   bvh.findPoints(offsets, counts, candidates, ncells, zip_test);
-
-  EXPECT_TRUE(candidates != nullptr);
 
   spin::UniformGrid<IndexType, NDIMS> ug(lo, hi, res);
 
@@ -1251,9 +1209,6 @@ void check_find_points_zip3d()
   axom::deallocate(xs);
   axom::deallocate(ys);
   axom::deallocate(zs);
-  axom::deallocate(offsets);
-  axom::deallocate(candidates);
-  axom::deallocate(counts);
   axom::deallocate(aabbs);
 
   axom::setDefaultAllocator(current_allocator);
@@ -1327,12 +1282,10 @@ void check_find_points_zip2d()
   }
 
   // traverse the BVH to find the candidates for all the centroids
-  IndexType* offsets = axom::allocate<IndexType>(ncells);
-  IndexType* counts = axom::allocate<IndexType>(ncells);
-  IndexType* candidates = nullptr;
-  bvh.findPoints(offsets, counts, candidates, ncells, centroids);
-
-  EXPECT_TRUE(candidates != nullptr);
+  axom::Array<IndexType> offsets(ncells);
+  axom::Array<IndexType> counts(ncells);
+  axom::Array<IndexType> candidates;
+  bvh.findPoints(offsets, counts, candidates, ncells, zip_test);
 
   spin::UniformGrid<IndexType, NDIMS> ug(lo, hi, res);
 
@@ -1344,9 +1297,6 @@ void check_find_points_zip2d()
     EXPECT_EQ(donorCellIdx, candidates[offsets[i]]);
   }  // END for all cell centroids
 
-  axom::deallocate(offsets);
-  axom::deallocate(candidates);
-  axom::deallocate(counts);
   axom::deallocate(aabbs);
 
   axom::setDefaultAllocator(current_allocator);
@@ -1677,20 +1627,17 @@ AXOM_CUDA_TEST(spin_bvh, use_pool_allocator)
       centroid[idx] = PointType {0.5, 0.5, 0.5};
     });
 
-  IndexType* offsets = axom::allocate<IndexType>(NUM_BOXES, allocID);
-  IndexType* counts = axom::allocate<IndexType>(NUM_BOXES, allocID);
-  IndexType* candidates = nullptr;
+  axom::Array<IndexType> offsets(NUM_BOXES, NUM_BOXES, allocID);
+  axom::Array<IndexType> counts(NUM_BOXES, NUM_BOXES, allocID);
+  axom::Array<IndexType> candidates;
   bvh.findPoints(offsets, counts, candidates, NUM_BOXES, centroid);
-  EXPECT_TRUE(candidates != nullptr);
+  EXPECT_TRUE(candidates.size() > 0);
 
   // Ensure the BVH uses interally the supplied pool allocator
-  EXPECT_EQ(rm.getAllocator(candidates).getId(), allocID);
+  EXPECT_EQ(candidates.getAllocatorID(), allocID);
 
   axom::deallocate(centroid);
   axom::deallocate(boxes);
-  axom::deallocate(offsets);
-  axom::deallocate(counts);
-  axom::deallocate(candidates);
 }
 
 #endif /* AXOM_USE_CUDA && AXOM_USE_RAJA && AXOM_USE_UMPIRE */